Horses or Humans/requirements-cpu.txt
fastapi
hydra-core
matplotlib
mlflow
numpy
pillow
pylint
python-json-logger
python-multipart
pyyaml
rich
ruff
torch
torchvision
tqdm
uvicorn